+++ src/motor/MotorDriver.h
@@ -33,8 +33,8 @@ struct VectorXY
 };
 
 struct SpeedRange {
-    int min;
-    int max;
+    float min;
+    float max;
 };
 
 class MotorDriver {
@@ -62,7 +62,7 @@ class MotorDriver {
 public:
     MotorDriver(const MotorPin m1, const MotorPin m2, const MotorPin m3) : m1(m1), m2(m2), m3(m3) {};
 
-    void init(int minSpeed = MIN_SPEED, int maxSpeed = MAX_SPEED);
+    void init(float minSpeed = MIN_SPEED, float maxSpeed = MAX_SPEED);
 
     void updateAllMotors() const;
     void syncUpdateAllMotors() const;
